Optimized Xsens Rx buffering
(Mostly) copy of commit of same name from mqtt-connect. Only differences are that buffer size is 224 (224 is smallest multiple of 32 that is greater than 200, not 256) and that cache invalidate doesn't invalidate entire buffer, just the cache lines corresponding to memory where new data is.

+++ TelemHost_Firmware_FE13/Core/Inc/serial.h
@@ -19,7 +19,7 @@ typedef struct
 	uint32_t index;
 } Serial_t;
 
-void Serial_Init(Serial_t* serial, UART_HandleTypeDef* h_uart, uint8_t* buf, uint8_t len);
+void Serial_Init(Serial_t* serial, UART_HandleTypeDef* h_uart, uint8_t* buf, uint16_t len);
 HAL_StatusTypeDef Serial_StartListening(Serial_t* serial);
 uint32_t Serial_BytesAvailable(Serial_t* serial);
 uint8_t Serial_GetByte(Serial_t* serial);
